Move policy code into wire.h so it can be shared with val.h and bind.h.
